// webpack/model/paths/IPaths.ts
export interface IPaths {
  Entry: string;
  Template: string;
  BuildDir: string;
  SrcDir: string;
  EngineDir: string;
  ModulesDir: string;
  ControllersDir: string;
  ExceptionsDir: string;
  InterfacesDir: string;
  MiddlewaresDir: string;
  ModelsDir: string;
  RoutesDir: string;
  ServicesDir: string;
}